backend/pyproject.toml
[project]
name = "vpasi-backend"
version = "0.1.0"
description = "vPasi — AI-powered trade assistant for informal cross-border traders in Africa"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "africastalking>=1.2.7",
    "google-cloud-aiplatform>=1.74.0",
    "langchain-google-genai>=2.1.0",
    "langchain-google-vertexai>=2.0.0",
    "langgraph>=0.2.0",
    "redis>=5.2.0",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.5",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "ddgs>=9.11.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-anyio>=0.0.0",
    "anyio>=4.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
